+++ Example/FeeRateKit/ViewController.swift
@@ -8,32 +8,34 @@ class ViewController: UIViewController {
     @IBOutlet weak var textView: UITextView?
 
     private let feeRateKit = Kit.instance(providerConfig: FeeProviderConfig(
-            infuraProjectId: "2a1306f1d12f4c109a4d4fb9be46b02e", 
-            infuraProjectSecret: "fc479a9290b64a84a15fa6544a130218", 
+            ethEvmUrl: FeeProviderConfig.infuraUrl(projectId: "2a1306f1d12f4c109a4d4fb9be46b02e"),
+            ethEvmAuth: "fc479a9290b64a84a15fa6544a130218",
+            bscEvmUrl: FeeProviderConfig.defaultBscEvmUrl,
             btcCoreRpcUrl: "https://btc.horizontalsystems.xyz/rpc",
             btcCoreRpcUser: nil,
             btcCoreRpcPassword: nil
     ))
 
-    private let exampleCoins = ["BTC", "LTC", "BCH", "DASH", "ETH"]
+    private let exampleBlockchains = ["BTC", "LTC", "BCH", "DASH", "ETH", "BSC"]
 
     override func viewDidLoad() {
         super.viewDidLoad()
     }
-    
-    func getRate(coin: String) -> Single<Int> {
-        switch coin {
+
+    func getRate(blockchain: String) -> Single<Int> {
+        switch blockchain {
             case "BTC": return feeRateKit.bitcoin(blockCount: 10)
             case "LTC": return feeRateKit.litecoin
             case "BCH": return feeRateKit.bitcoinCash
             case "DASH": return feeRateKit.dash
             case "ETH": return feeRateKit.ethereum
+            case "BSC": return feeRateKit.binanceSmartChain
             default: return Single.just(0)
         }
     }
 
     @IBAction func refresh() {
-        Single.zip(exampleCoins.map { getRate(coin: $0) })
+        Single.zip(exampleBlockchains.map { getRate(blockchain: $0) })
                 .subscribeOn(ConcurrentDispatchQueueScheduler(qos: .userInitiated))
                 .observeOn(MainScheduler.instance)
                 .subscribe(onSuccess: { [weak self] feeRates in
@@ -48,20 +50,21 @@ class ViewController: UIViewController {
         var ratesString = ""
         
         for (index, rate) in rates.enumerated() {
-            ratesString += "[\(name(from: exampleCoins[index]))]\nRate: \(rate)\n\n"
+            ratesString += "[\(name(from: exampleBlockchains[index]))]\nRate: \(rate)\n\n"
         }
         
         textView?.text = ratesString
     }
 
-    private func name(from code: String) -> String {
-        switch code {
+    private func name(from blockchain: String) -> String {
+        switch blockchain {
         case "BTC": return "Bitcoin"
         case "LTC": return "Litecoin"
         case "ETH": return "Ethereum"
         case "DASH": return "Dash"
         case "BCH": return "Bitcoin Cash"
-        default: return "Unknown" 
+        case "BSC": return "Binance Smart Chain"
+        default: return "Unknown"
         }
     }