+++ CHANGELOG.md
@@ -3,6 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## Current Version
 
+* Add `Binance Smart Chain` support, change structure of `FeeProviderConfig` [ **non-back-compatible api change** ]
+
 ## 0.6.0
 
 * Remove FeeRate class. Remove cache. Fixed rates for LTC, DASH and BCH [ **non-back-compatible api change** ]